+++ packages/terra-framework-docs/CHANGELOG.md
@@ -5,6 +5,13 @@
 
 ## Unreleased
 
+## 1.7.0 - (March 8, 2022)
+
+* Added
+  * Migrate docs from terra-layout
+  * Added docs for `terra-filter-pills`
+  * Added tests for `terra-filter-pills`
+
 ## 1.6.0 - (November 16, 2021)
 
 * Changed